requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
yookassa==3.3.0
python-dotenv==1.0.1
